+++ Circus Of Plates/src/model/SpriteShape.java
@@ -7,8 +7,8 @@ public class SpriteShape extends Sprite {
 
 	private int width, height;
 	private Color color;
-	
-	public SpriteShape(int x, int y, int width, int height, Color color) {
+
+	public SpriteShape(double x, double y, int width, int height, Color color) {
 		super(x, y);
 		this.width = width;
 		this.height = height;
